requirements.txt
Django
MySQL-python
Sphinx
Werkzeug
django-extensions
requests
furl
ordereddict
jsonpickle
M2Crypto
pycrypto
enum34
